requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
dominate==2.8.0
einops==0.7.0
faiss-cpu==1.7.4
fast-pytorch-kmeans==0.2.0.1
filelock==3.12.4
fonttools==4.44.0
frozenlist==1.4.1
fsspec==2023.10.0
gdown==4.7.1
h5py==3.10.0
huggingface-hub==0.21.4
idna==3.4
imageio==2.31.5
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
loralib==0.1.2
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
peft==0.10.0
Pillow==10.1.0
prettytable==3.10.0
psutil==5.9.8
pynvml==11.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==2.2.1
pytorch-metric-learning==2.4.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
soupsieve==2.5
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.15.2
torch==2.2.0+cu121
torch_geometric==2.5.2
torchmetrics==1.3.1
torchvision==0.17.0+cu121
tqdm==4.66.1
transformers==4.39.0
triton==2.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.13
xformers==0.0.24
yarl==1.9.4
zipp==3.17.0
